Reformat src/types
